Refactor DuplicateApplication into two separate services
